backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
pydantic-settings==2.2.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
bcrypt==4.1.2
pytest==8.0.1
pytest-asyncio==0.23.5
